requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
idna==3.4
importlib-metadata==6.5.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
lxml==4.9.3
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pygame==2.4.0
Pygments==2.15.1
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.0.2
qrcode==7.4.2
requests==2.31.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
soupsieve==2.5
stack-data==0.6.2
tornado==6.3
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==2.0.6
wcwidth==0.2.6
zipp==3.15.0
